- correct casting float -> double for SIMD - non-SIMD code uses double precision as default

+++ MapProjections/GeoCoordinate.h
@@ -1,7 +1,7 @@
 #ifndef GEOCORDINATE_H
 #define GEOCORDINATE_H
 
-typedef float MyRealType;
+typedef double MyRealType;
 
 template <typename T>
 struct IAngle
